// plugins/tinker-shortcut/src/renderer/types.ts
export type ActionType = 'command' | 'app' | 'directory' | 'url' | 'plugin'

export type FilterTab = 'all' | ActionType

export interface Action {
  id: string
  name: string
  shortcut: string
  enabled: boolean
  type: ActionType
  command: string
  appIcon?: string
}

export type ActionInput = Omit<Action, 'id'>